Extract meta information for books

find_book() extracts meta-data from JSTOR-XML files for book chapters.

find_book(file_path)

Arguments

file_path
A .xml-file for a book or research report.

Value

A tibble containing the extracted meta-data with the following columns:

basename_id (chr): The filename of the original .xml-file. Can be used for joining with other data for the same file.

discipline (chr): The discipline from the discipline names used on JSTOR.

book_id (chr): The book id of type "jstor", which is not a registered DOI.

book_title (chr): The title of the book.

book_subtitle (chr): The subtitle of the book.

pub_day (int): Publication day, if specified.

pub_month (int): Publication month, if specified.

pub_year (int): Year of publication.

isbn (chr): One or more entries for the book's ISBN. If two or more, separated by "; ".

publisher_name (int): The name of the publisher.

publisher_loc (int): The location of the publisher.

n_pages (int): The number of pages.

language (chr): The language of the book.

A note about publication dates: always the first entry is being extracted, which should correspond to the oldest date, in case there is more than one date.

Examples

find_book(jstor_example("sample_book.xml"))

#> # A tibble: 1 x 13
#>   book_id  basename_id discipline  book_title   book_subtitle pub_day pub_month
#>   <chr>    <chr>       <chr>       <chr>        <chr>           <int>     <int>
#> 1 j.ctt24… sample_book Political … The 2006 Mi… A Coup to En…      30         4
#> # ... with 6 more variables: pub_year <int>, isbn <chr>, publisher_name <chr>,
#> #   publisher_location <chr>, n_pages <int>, language <chr>