Extract all footnotes

This function extracts the content of fn-group from journal-articles.

jst_get_footnotes(file_path)

Arguments

file_path
The path to the .xml-file from which footnotes should be extracted.

Value

A tibble containing the content from fn-group (usually the footnotes). If there were no footnotes, NA_character is returned for the column footnotes.

Details

The fn-group usually contains footnotes corresponding to the article. However, since footnotes are currently not fully supported by DfR, there is no comprehensive documentation on the different variants. jstor therefore extracts the content of fn-group exactly as it appears in the data. Because of this, there might be other content present than footnotes.

In order to get all available information on citation data, you might need to combine jst_get_footnotes() with jst_get_references().

Examples

jst_get_footnotes(jst_example("article_with_footnotes.xml"))

#> # A tibble: 7 x 2
#>   file_name            footnotes                                                
#>   <chr>                <chr>                                                    
#> 1 article_with_footno… "[Footnotes]"                                            
#> 2 article_with_footno… "9\nQuarterly, vol. XIII, no. 1,\nentries for April 19 a…
#> 3 article_with_footno… "10\nQuarterly, vol. XIII,\nno. 1, p. 8."                
#> 4 article_with_footno… "14\nQuarterly, vol. VIII, no. 1.\nOlympia Columbian, Se…
#> 5 article_with_footno… "26\nQuarterly, vol. XII,\nno. 2, p. 141."               
#> 6 article_with_footno… "32\nDr. David S. Maynard, later (March 31, 1852)"       
#> 7 article_with_footno… "34\nThomas Linklater, Shepherd, since October 6, 1849,"