Re-import files

jst_re_import() lets you re-import a file which was exported via jstor_import() or jst_import_zip().

jst_re_import(file, warn = TRUE)

Arguments

file
A path to a .csv file.

warn
Should warnings be emitted, if the type of file cannot be determined?

Details

When attempting to re-import, a heuristic is applied. If the file has column names which match the names from any of the find_* functions, the file is read with the corresponding specifications. If no column names are recognized, files are recognized based on the number of columns. Since both references and footnotes have only two columns, the first line is inspected for either "Referenc...|Bilbio...|Endnote..." or "Footnote...". In case there is still no match, the file is read with readr::read_csv() with guess_max = 5000 and a warning is raised.

See also