Defunct: Wrapper for file import

This function will be removed in the next release. Please use jst_import() instead.

jstor_import(in_paths, out_file, out_path = NULL, .f, col_names = TRUE,
  n_batches = NULL, files_per_batch = NULL,
  cores = getOption("mc.cores", 1L), show_progress = TRUE)

Arguments

in_paths
A character vector to the xml-files which should be imported

out_file
Name of files to export to. Each batch gets appended by an increasing number.

out_path
Path to export files to (combined with filename).

.f
Function to use for import. Can be one of find_article, find_authors, find_references, find_footnotes, find_book or find_chapter.

col_names
Should column names be written to file? Defaults to TRUE.

n_batches
Number of batches, defaults to 1.

files_per_batch
Number of files for each batch. Can be used instead of n_batches, but not in conjunction.

cores
Number of cores to use for parallel processing.

show_progress
Displays a progress bar for each batch, if the session is interactive.

Value

Writes .csv-files to disk.

Details

This function applies an import function to a list of xml-files or a .zip-archive in case of jstor_import_zip and saves the output in batches of .csv-files to disk.

Along the way, we wrap three functions, which make the process of converting many files easier:

purrr::safely()

foreach::foreach()

readr::write_csv()

When using one of the find_* functions, there should usually be no errors. To avoid the whole computation to fail in the unlikely event that an error occurs, we use safely() which let's us continue the process, and catch the error along the way.

If you have many files to import, you might benefit from executing the function in parallel. We use snow::createCluster() to setup a cluster and then compute the results via foreach and %dopar%. The type of cluster is determined by getClusterOption("type").

After importing all files, they are written to disk with readr::write_csv().

Since you might run out of memory when importing a large quantity of files, the files to import are split up into batches. Each batch is being treated separately, therefore for each batch multiple processes from snow::createCluster() are spawned. For this reason, it is not recommended to have very small batches, as there is an overhead for starting and ending the processes. On the other hand, the batches should not be too large, to not exceed memory limitations. A value of 10000 to 20000 for files_per_batch should work fine on most machines. If the session is interactive and show_progress is TRUE, a progress bar is displayed for each batch.

See also

Examples

# NOT RUN {
# read from file list --------
# find all files
meta_files <- list.files(pattern = "xml", full.names = T)

# import them via `find_article`
jstor_import(meta_files, out_file = "imported_metadata", .f = find_article,
             files_per_batch = 25000, cores = 4)

# read from zip archive ------ 
# define imports
imports <- jst_define_import(article = c(find_article, find_authors))

# convert the files to .csv
jstor_import_zip("my_archive.zip", out_file = "my_out_file",
                 import_spec = imports)
# }