Get table with information on journals

Download most recent or display cached version of data on journals.

jst_get_journal_overview(most_recent = FALSE)

Arguments

most_recent
Should the most recent version be downloaded from DfR?

Value

A tibble with various information about journals.

Details

When analysing your sample of articles from DfR, it might be helpful to have some context about the journals in your sample. This function provides a tibble with various information like the full name of the journal, the short version of the name (sometimes referred to as JCODE), dates on where the first and last (available) issues were published, etc.

The data on journals might change. Therefore this function provides two sources of data: a cached version which gets updated with every release, and the ability to pull the most recent version directly from DfR.

The cached version was updated on 2018-05-15.

Examples

# use the function without arguments to get a tibble from disk
jst_get_journal_overview()

#> # A tibble: 3,709 x 29
#>    publication_title        print_identifier online_identifi… date_first_issue…
#>    <chr>                    <chr>            <chr>            <date>           
#>  1 Acta Classica            0065-1141        2227-538X        1958-01-01       
#>  2 Acta Criminologica       0065-1168        1718-3243        1968-01-01       
#>  3 Acta Historica Academia… 0001-5849        2064-7182        1951-01-01       
#>  4 Acta Linguistica Academ… 0001-5946        2560-1814        1951-01-01       
#>  5 Acta Linguistica Hungar… 1216-8076        1588-2624        1988-01-01       
#>  6 Acta Musicologica        0001-6241        <NA>             1931-01-01       
#>  7 Acta Oeconomica          0001-6373        1588-2659        1966-01-01       
#>  8 Acta Orientalia Academi… 0001-6446        1588-2667        1950-01-01       
#>  9 Acta Sociologica         0001-6993        <NA>             1955-01-01       
#> 10 Acta Turistica           0353-4316        1848-6061        1989-01-01       
#> # ... with 3,699 more rows, and 25 more variables: num_first_vol_online <chr>,
#> #   num_first_issue_online <chr>, date_last_issue_online <date>,
#> #   num_last_vol_online <chr>, num_last_issue_online <chr>, title_url <chr>,
#> #   title_id <chr>, embargo_info <chr>, coverage_depth <chr>, notes <chr>,
#> #   publisher_name <chr>, publication_type <chr>,
#> #   date_monograph_published_print <date>,
#> #   date_monograph_published_online <date>, monograph_volume <chr>,
#> #   monograph_edition <chr>, parent_publication_title_id <chr>,
#> #   preceding_publication_title_id <chr>, access_type <chr>,
#> #   full_coverage <chr>, collection <chr>, discipline <chr>,
#> #   catalog_identifier_oclc <chr>, catalog_identifier_lccn <chr>,
#> #   archive_release_date <date>

# NOT RUN {
# download the most recent version from DfR
jst_get_journal_overview(most_recent = TRUE)
# }