Get table with information on journals

Download most recent or display cached version of data on journals.

jst_get_journal_overview(most_recent = FALSE, quiet = FALSE)

Arguments

most_recent
Should the most recent version be downloaded from DfR?

quiet
Should status messages about the download be printed?

Value

A tibble with various information about journals.

Details

When analysing your sample of articles from DfR, it might be helpful to have some context about the journals in your sample. This function provides a tibble with various information like the full name of the journal, the short version of the name (sometimes referred to as JCODE), dates on where the first and last (available) issues were published, etc.

The data on journals might change. Therefore this function provides two sources of data: a cached version which gets updated with every release, and the ability to pull the most recent version directly from DfR.

The cached version was updated on 2018-05-15.

Examples

# use the function without arguments to get a tibble from disk
jst_get_journal_overview()

#> # A tibble: 4,132 x 12
#>    title journal_id issn  eissn doi   url   discipline publisher coverage_range
#>    <chr> <chr>      <chr> <chr> <chr> <chr> <chr>      <chr>     <chr>         
#>  1 14th… 14centeng… 0737… <NA>  10.2… http… Humanitie… Penn Sta… 1974-1983     
#>  2 19th… 19thcentu… 0148… 1533… 10.2… http… Arts ; Mu… Universi… 1977-2014     
#>  3 291   291        1054… <NA>  10.2… http… Art & Art… Thomas J… 1915-1916     
#>  4 4S R… 4sreview   0738… <NA>  10.2… http… History ;… Sage Pub… 1972-1985     
#>  5 A.A.… aavnewsle… 2154… <NA>  10.2… http… Biologica… Associat… 1980-1986     
#>  6 A.I.… aihpnotes  2161… 2329… 10.2… http… History ;… American… 1955-1957     
#>  7 AA F… aafiles    0261… <NA>  10.2… http… Architect… Architec… 1981-2017     
#>  8 AAA:… aaaarbean… 0171… <NA>  10.2… http… American … Narr Fra… 1976-2016     
#>  9 AATS… aatseeljo… 2325… 2379… 10.2… http… Area Stud… American… 1945-1956     
#> 10 AAUP… aaupbulle… 0001… <NA>  10.2… http… Education… American… 1915-1978     
#> # ... with 4,122 more rows, and 3 more variables:
#> #   oclc_catalog_identifier <chr>, lccn_catalog_identifier <chr>,
#> #   archive_release_date <chr>

# NOT RUN {
# download the most recent version from DfR
jst_get_journal_overview(most_recent = TRUE)
# }