Extract all references

This function extracts the content of ref-list from the xml-file.

jst_get_references(file_path, parse_refs = F)

Arguments

file_path
The path to the .xml-file from which references should be extracted.

parse_refs
Should references be parsed, if available?

Value

A tibble with the following columns:

file_name: the identifier for the article the references come from.

ref_title: the title of the references sections.

ref_authors: a string of authors. Several authors are separated with ;.

ref_editors: a string of editors, if available.

ref_collab: a field that may contain information on the authors, if authors are not available.

ref_item_title: the title of the cited entry. For books this is often empty, with the title being in ref_source.

ref_year: a year, often the article's publication year, but not always.

ref_source: the source of the cited entry. For books often the title of the book, for articles the publisher of the journal.

ref_volume: the volume of the journal article.

ref_first_page: the first page of the article/chapter.

ref_last_page: the last page of the article/chapter.

ref_publisher: For books the publisher, for articles often missing.

ref_publication_type: Known types: book, journal, web, other.

ref_unparsed: The full references entry in unparsed form.

Details

This content may contain references or endnotes, depending on how the article used citations. Since references are currently not fully supported by DfR, there is no comprehensive documentation on the different variants. jstor therefore extracts the content of ref-list exactly as it appears in the data. Because of this, there might be other content present than references.

In order to get all available information on citation data, you might need to combine jst_get_references() with jst_get_footnotes().

For newer xml-files, there would be the option to extract single elements like authors, title or date of the source, but this is not yet implemented.

In general, the implementation is not as fast as jst_get_article() - articles with many references slow the process down.

Examples

jst_get_references(jst_example("article_with_references.xml"))

#> # A tibble: 23 x 13
#>    file_name ref_title ref_authors ref_collab ref_item_title ref_year ref_source
#>    <chr>     <chr>     <chr>       <chr>      <chr>          <chr>    <chr>     
#>  1 article_… Bibliogr… NA          NA         NA             NA       NA        
#>  2 article_… Bibliogr… NA          NA         NA             NA       NA        
#>  3 article_… Referenc… NA          NA         NA             NA       NA        
#>  4 article_… Referenc… NA          NA         NA             NA       NA        
#>  5 article_… Referenc… NA          NA         NA             NA       NA        
#>  6 article_… Referenc… NA          NA         NA             NA       NA        
#>  7 article_… Referenc… NA          NA         NA             NA       NA        
#>  8 article_… Referenc… NA          NA         NA             NA       NA        
#>  9 article_… Referenc… NA          NA         NA             NA       NA        
#> 10 article_… Referenc… NA          NA         NA             NA       NA        
#> # ... with 13 more rows, and 6 more variables: ref_volume <chr>,
#> #   ref_first_page <chr>, ref_last_page <chr>, ref_publisher <chr>,
#> #   ref_publication_type <chr>, ref_unparsed <chr>

# import parsed references
jst_get_references(
  jst_example("parsed_references.xml"),
  parse_refs = TRUE
)

#> # A tibble: 7 x 14
#>   file_name ref_title ref_authors ref_editors ref_collab ref_item_title ref_year
#>   <chr>     <chr>     <chr>       <chr>       <chr>      <chr>          <chr>   
#> 1 parsed_r… Notes     NA          NA          NA         NA             2005    
#> 2 parsed_r… Referenc… Acohido, B… NA          NA         “Snowden Case… 2013    
#> 3 parsed_r… Referenc… NA          NA          Amnesty I… NA             2013    
#> 4 parsed_r… Referenc… Jacobson, … D. E. Davi… NA         Chapter title  2009    
#> 5 parsed_r… Referenc… Costall, A… NA          NA         “Some article… 1980    
#> 6 parsed_r… Referenc… Hudson, W.  NA          NA         Another artic… 2000    
#> 7 parsed_r… Referenc… Fries-Brit… NA          NA         Some article … 2000    
#> # ... with 7 more variables: ref_source <chr>, ref_volume <chr>,
#> #   ref_first_page <chr>, ref_last_page <chr>, ref_publisher <chr>,
#> #   ref_publication_type <chr>, ref_unparsed <chr>