Extract all references

This function extracts the content of ref-list from the xml-file.

jst_get_references(file_path, parse_refs = F)

Arguments

file_path
The path to the .xml-file from which references should be extracted.

parse_refs
Should references be parsed, if available?

Value

A tibble with the following columns:

file_name: the identifier for the article the references come from.

ref_title: the title of the references sections.

authors: a string of authors. Several authors are separated with ;.

editors: a string of editors, if available.

collab: a field that may contain information on the authors, if authors are not available.

title: the title of the cited entry. For books this is often empty, with the title being in source.

year: a year, often the article's publication year, but not always.

source: the source of the cited entry. For books often the title of the book, for articles the publisher of the journal.

volume: the volume of the journal article.

first_page: the first page of the article/chapter.

last_page: the last page of the article/chapter.

publisher: For books the publisher, for articles often missing.

publication_type: Known types: book, journal, web, other.

unparsed_refs: The full references entry in unparsed form.

Details

This content may contain references or endnotes, depending on how the article used citations. Since references are currently not fully supported by DfR, there is no comprehensive documentation on the different variants. jstor therefore extracts the content of ref-list exactly as it appears in the data. Because of this, there might be other content present than references.

In order to get all available information on citation data, you might need to combine jst_get_references() with jst_get_footnotes().

For newer xml-files, there would be the option to extract single elements like authors, title or date of the source, but this is not yet implemented.

In general, the implementation is not as fast as jst_get_article() - articles with many references slow the process down.

Examples

jst_get_references(jst_example("article_with_references.xml"))

#> # A tibble: 23 x 13
#>    file_name ref_title authors collab title year  source volume first_page
#>    <chr>     <chr>     <chr>   <chr>  <chr> <chr> <chr>  <chr>  <chr>     
#>  1 article_… Bibliogr… <NA>    <NA>   <NA>  <NA>  <NA>   <NA>   <NA>      
#>  2 article_… Bibliogr… <NA>    <NA>   <NA>  <NA>  <NA>   <NA>   <NA>      
#>  3 article_… Referenc… <NA>    <NA>   <NA>  <NA>  <NA>   <NA>   <NA>      
#>  4 article_… Referenc… <NA>    <NA>   <NA>  <NA>  <NA>   <NA>   <NA>      
#>  5 article_… Referenc… <NA>    <NA>   <NA>  <NA>  <NA>   <NA>   <NA>      
#>  6 article_… Referenc… <NA>    <NA>   <NA>  <NA>  <NA>   <NA>   <NA>      
#>  7 article_… Referenc… <NA>    <NA>   <NA>  <NA>  <NA>   <NA>   <NA>      
#>  8 article_… Referenc… <NA>    <NA>   <NA>  <NA>  <NA>   <NA>   <NA>      
#>  9 article_… Referenc… <NA>    <NA>   <NA>  <NA>  <NA>   <NA>   <NA>      
#> 10 article_… Referenc… <NA>    <NA>   <NA>  <NA>  <NA>   <NA>   <NA>      
#> # ... with 13 more rows, and 4 more variables: last_page <chr>,
#> #   publisher <chr>, publication_type <chr>, unparsed_refs <chr>

# import parsed references
jst_get_references(
  jst_example("parsed_references.xml"),
  parse_refs = TRUE
)

#> # A tibble: 7 x 14
#>   file_name ref_title authors editors collab title year  source volume
#>   <chr>     <chr>     <chr>   <chr>   <chr>  <chr> <chr> <chr>  <chr> 
#> 1 parsed_r… Notes     <NA>    <NA>    <NA>   <NA>  2005  <NA>   <NA>  
#> 2 parsed_r… Referenc… Acohid… <NA>    <NA>   “Sno… 2013  USA T… <NA>  
#> 3 parsed_r… Referenc… <NA>    <NA>    Amnes… <NA>  2013  “USA:… <NA>  
#> 4 parsed_r… Referenc… Jacobs… D. E. … <NA>   Chap… 2009  Book … <NA>  
#> 5 parsed_r… Referenc… Costal… <NA>    <NA>   “Som… 1980  Theor… 1     
#> 6 parsed_r… Referenc… Hudson… <NA>    <NA>   Anot… 2000  Austr… 40    
#> 7 parsed_r… Referenc… Fries-… <NA>    <NA>   Some… 2000  Journ… 20    
#> # ... with 5 more variables: first_page <chr>, last_page <chr>,
#> #   publisher <chr>, publication_type <chr>, unparsed_refs <chr>