Extract all references

This function extracts the content of ref-list from the xml-file.

jst_get_references(file_path)

Arguments

file_path
The path to the .xml-file from which references should be extracted.

Value

A tibble with three two containing the references:

basename_id: the identifier for the article the references come from.

references: the text of the references.

Details

This content may contain references or endnotes, depending on how the article used citations. Since references are currently not fully supported by DfR, there is no comprehensive documentation on the different variants. jstor therefore extracts the content of ref-list exactly as it appears in the data. Because of this, there might be other content present than references.

In order to get all available information on citation data, you might need to combine jst_get_references() with jst_get_footnotes().

For newer xml-files, there would be the option to extract single elements like authors, title or date of the source, but this is not yet implemented.

In general, the implementation is not as fast as jst_get_article() - articles with many references slow the process down.

Examples

jst_get_references(jst_example("sample_with_references.xml"))

#> # A tibble: 29 x 2
#>    file_name           references                                              
#>    <chr>               <chr>                                                   
#>  1 sample_with_refere… Bibliography: Entamoeba ranarumn                        
#>  2 sample_with_refere… "DOBELL, C.C.\n1909 Researches on the intestinal Protoz…
#>  3 sample_with_refere… "1918 Are Entamoeba histolytica and Entamoeba ranarum t…
#>  4 sample_with_refere… References: Leptotheca ohilmacheri                      
#>  5 sample_with_refere… "KUDO, R.\n1920 Studies on Myxosporidia. A Synopsis of …
#>  6 sample_with_refere… "1921 On the nature of structures characteristic of Cni…
#>  7 sample_with_refere… "1922 On the morphology and life history of a Myxospori…
#>  8 sample_with_refere… References: Haemogregarines                             
#>  9 sample_with_refere… "DUTTON, J. E., J. L. TODD and E. N. TOBEY.\n1907 Conce…
#> 10 sample_with_refere… "FRANNÇA, C.\n1917 Sur la classification des hémosporid…
#> # ... with 19 more rows